+++ pytest_doctestplus/plugin.py
@@ -98,12 +98,6 @@ def pytest_addoption(parser):
                          "This is no longer recommended, use --doctest-glob instead."
                      ))
 
-    parser.addoption("--doctest-ufunc", action="store_true",
-                     help=(
-                         "Enable running doctests in docstrings of Numpy ufuncs. "
-                         "Implies usage of doctest-plus."
-                     ))
-
     # Defaults to `atol` parameter from `numpy.allclose`.
     parser.addoption("--doctest-plus-atol", action="store",
                      help="set the absolute tolerance for float comparison",
@@ -133,10 +127,6 @@ def pytest_addoption(parser):
 
     parser.addini("doctest_rst",
                   "Run the doctests in the rst documentation",
-                  default=False)
-
-    parser.addini("doctest_ufunc",
-                  "Run doctests in docstrings of Numpy ufuncs",
                   default=False)
 
     parser.addini("doctest_plus_atol",
@@ -184,9 +174,8 @@ def _is_numpy_ufunc(method):
 def pytest_configure(config):
     doctest_plugin = config.pluginmanager.getplugin('doctest')
     run_regular_doctest = config.option.doctestmodules and not config.option.doctest_plus
-    use_ufunc = config.getini('doctest_ufunc') or config.option.doctest_ufunc
     use_doctest_plus = config.getini(
-        'doctest_plus') or config.option.doctest_plus or config.option.doctest_only or use_ufunc
+        'doctest_plus') or config.option.doctest_plus or config.option.doctest_only
     if doctest_plugin is None or run_regular_doctest or not use_doctest_plus:
         return
 
@@ -264,11 +253,10 @@ def pytest_configure(config):
                 verbose=False, optionflags=options, checker=OutputChecker())
 
             tests = finder.find(module)
-            if use_ufunc:
-                for method in module.__dict__.values():
-                    if _is_numpy_ufunc(method):
-                        found = finder.find(method, module=module)
-                        tests += found
+            for method in module.__dict__.values():
+                if _is_numpy_ufunc(method):
+                    found = finder.find(method, module=module)
+                    tests += found
 
             for test in tests:
                 if test.examples:  # skip empty doctests